Parse doubles to a new Double class rather than to the Float class.
This is simply a case of not discarding information. Without this, it would
not be possible to parse a trace and re-create the identical trace.

+++ common/trace_parser.cpp
@@ -551,7 +551,7 @@ void Parser::scan_float() {
 Value *Parser::parse_double() {
     double value;
     file->read(&value, sizeof value);
-    return new Float(value);
+    return new Double(value);
 }
